requirements.txt
pandas==1.5.0
numpy==1.23.3
plotly==5.10.0
requests==2.28.1
transformers==4.22.1
python-dotenv==0.21.0
sentencepiece==0.1.97